Configured org-gcal for syncing calendars with google
I have created a new org-caldav calendar to store there the events
that I want to sync from org mode and the google calendar. It is
importnt to note that events cannot have child or lists, just siple
events. So this is very limited.

I also tried org-caldav, is newer, but never worked, the
authentication was apparently solved by using plstore but still did
not sync.